backend/pyproject.toml
[project]
name = "rulebook-arbiter-backend"
version = "0.1.0"
description = "Board game rulebook Q&A backend powered by RAG"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "uvicorn",
    "pypdf",
    "chromadb",
    "google-genai",
    "pydantic-settings",
    "tenacity",
    "python-multipart",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "pytest-cov>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
